# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Tawsileh Intercity MVP API
 * OpenAPI spec version: 1.0.0
 */
export type CaptainVehicle = {
  makeModel?: string;
  plate?: string;
  color?: string;
  totalSeats?: number;
};

export interface Captain {
  id: string;
  status: string;
  isOnline: boolean;
  rating?: number;
  totalTrips?: number;
  vehicle?: CaptainVehicle;
}

export interface User {
  id: string;
  phone: string;
  fullName: string;
  role: string;
  captainProfile?: Captain;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface MatchingCaptain {
  captainId: string;
  fullName: string;
  rating: number;
  vehicleMakeModel: string;
  vehiclePlate: string;
  availableSeats: number;
  estimatedFare: number;
  distanceKm: number;
  etaMinutes: number;
  lat?: number;
  lng?: number;
}

export type RideRequestPassenger = {
  fullName?: string;
  phone?: string;
};

export type RideRequestCaptain = {
  fullName?: string;
  phone?: string;
  vehicleMakeModel?: string;
  vehiclePlate?: string;
};

export interface RideRequest {
  id: string;
  passengerId: string;
  matchedCaptainId?: string;
  routeId: string;
  status: string;
  pickupLat?: number;
  pickupLng?: number;
  destLat?: number;
  destLng?: number;
  seatsRequested?: number;
  estimatedFare: number;
  createdAt: string;
  passenger?: RideRequestPassenger;
  captain?: RideRequestCaptain;
}

export interface Route {
  id: string;
  name: string;
  origin: string;
  destination: string;
  baseFare: number;
}

export type RegisterBodyRole = typeof RegisterBodyRole[keyof typeof RegisterBodyRole];


export const RegisterBodyRole = {
  passenger: 'passenger',
  captain: 'captain',
} as const;

export type RegisterBody = {
  phone: string;
  password: string;
  fullName: string;
  role: RegisterBodyRole;
};

export type LoginBody = {
  phone: string;
  password: string;
};

export type SetupCaptainProfileBody = {
  nationalIdLast4: string;
  vehicleMakeModel: string;
  vehicleColor: string;
  vehiclePlate: string;
  totalSeats: number;
};

export type UpdateCaptainStatusBody = {
  isOnline?: boolean;
  routeId?: string;
  pickupRadiusKm?: number;
};

export type UpdateCaptainLocationBody = {
  lat: number;
  lng: number;
};

export type SearchMatchingCaptainsParams = {
routeId: string;
pickupLat: number;
pickupLng: number;
destLat: number;
destLng: number;
seats: number;
};

export type CreateRideRequestBody = {
  captainId: string;
  routeId: string;
  pickupLat: number;
  pickupLng: number;
  destLat: number;
  destLng: number;
  seats: number;
};

export type ListRideRequestsParams = {
role?: ListRideRequestsRole;
status?: string;
};

export type ListRideRequestsRole = typeof ListRideRequestsRole[keyof typeof ListRideRequestsRole];


export const ListRideRequestsRole = {
  passenger: 'passenger',
  captain: 'captain',
} as const;

export type UpdateRideRequestStatusBodyStatus = typeof UpdateRideRequestStatusBodyStatus[keyof typeof UpdateRideRequestStatusBodyStatus];


export const UpdateRideRequestStatusBodyStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
  captain_arriving: 'captain_arriving',
  picked_up: 'picked_up',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type UpdateRideRequestStatusBody = {
  status: UpdateRideRequestStatusBodyStatus;
};