# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Tawsileh Intercity MVP API
servers:
  - url: /api
tags:
  - name: auth
  - name: captains
  - name: ride-requests
  - name: users
paths:
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password, fullName, role]
              properties:
                phone:
                  type: string
                password:
                  type: string
                fullName:
                  type: string
                role:
                  type: string
                  enum: [passenger, captain]
      responses:
        "200":
          description: Registered and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password]
              properties:
                phone:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /captains/profile:
    post:
      operationId: setupCaptainProfile
      tags: [captains]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nationalIdLast4, vehicleMakeModel, vehicleColor, vehiclePlate, totalSeats]
              properties:
                nationalIdLast4:
                  type: string
                vehicleMakeModel:
                  type: string
                vehicleColor:
                  type: string
                vehiclePlate:
                  type: string
                totalSeats:
                  type: number
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Captain"

  /captains/status:
    put:
      operationId: updateCaptainStatus
      tags: [captains]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isOnline:
                  type: boolean
                routeId:
                  type: string
                pickupRadiusKm:
                  type: number
      responses:
        "200":
          description: Status updated

  /captains/location:
    put:
      operationId: updateCaptainLocation
      tags: [captains]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                lng:
                  type: number
      responses:
        "200":
          description: Location updated

  /captains/matching:
    get:
      operationId: searchMatchingCaptains
      tags: [captains]
      parameters:
        - name: routeId
          in: query
          required: true
          schema:
            type: string
        - name: pickupLat
          in: query
          required: true
          schema:
            type: number
        - name: pickupLng
          in: query
          required: true
          schema:
            type: number
        - name: destLat
          in: query
          required: true
          schema:
            type: number
        - name: destLng
          in: query
          required: true
          schema:
            type: number
        - name: seats
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of matching captains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchingCaptain"

  /ride-requests:
    post:
      operationId: createRideRequest
      tags: [ride-requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [captainId, routeId, pickupLat, pickupLng, destLat, destLng, seats]
              properties:
                captainId:
                  type: string
                routeId:
                  type: string
                pickupLat:
                  type: number
                pickupLng:
                  type: number
                destLat:
                  type: number
                destLng:
                  type: number
                seats:
                  type: number
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideRequest"
    get:
      operationId: listRideRequests
      tags: [ride-requests]
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [passenger, captain]
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideRequest"

  /ride-requests/{id}/status:
    put:
      operationId: updateRideRequestStatus
      tags: [ride-requests]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [accepted, rejected, captain_arriving, picked_up, in_progress, completed, cancelled]
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideRequest"

  /routes:
    get:
      operationId: listRoutes
      tags: [operations]
      responses:
        "200":
          description: Available routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"

components:
  schemas:
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, phone, fullName, role]
      properties:
        id:
          type: string
        phone:
          type: string
        fullName:
          type: string
        role:
          type: string
        captainProfile:
          $ref: "#/components/schemas/Captain"
    Captain:
      type: object
      required: [id, status, isOnline]
      properties:
        id:
          type: string
        status:
          type: string
        isOnline:
          type: boolean
        rating:
          type: number
        totalTrips:
          type: number
        vehicle:
          type: object
          properties:
            makeModel:
              type: string
            plate:
              type: string
            color:
              type: string
            totalSeats:
              type: number
    MatchingCaptain:
      type: object
      required: [captainId, fullName, rating, vehicleMakeModel, vehiclePlate, availableSeats, estimatedFare, distanceKm, etaMinutes]
      properties:
        captainId:
          type: string
        fullName:
          type: string
        rating:
          type: number
        vehicleMakeModel:
          type: string
        vehiclePlate:
          type: string
        availableSeats:
          type: number
        estimatedFare:
          type: number
        distanceKm:
          type: number
        etaMinutes:
          type: number
        lat:
          type: number
        lng:
          type: number
    RideRequest:
      type: object
      required: [id, passengerId, routeId, status, estimatedFare, createdAt]
      properties:
        id:
          type: string
        passengerId:
          type: string
        matchedCaptainId:
          type: string
        routeId:
          type: string
        status:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        destLat:
          type: number
        destLng:
          type: number
        seatsRequested:
          type: number
        estimatedFare:
          type: number
        createdAt:
          type: string
          format: date-time
        passenger:
          type: object
          properties:
            fullName:
              type: string
            phone:
              type: string
        captain:
          type: object
          properties:
            fullName:
              type: string
            phone:
              type: string
            vehicleMakeModel:
              type: string
            vehiclePlate:
              type: string
    Route:
      type: object
      required: [id, name, origin, destination, baseFare]
      properties:
        id:
          type: string
        name:
          type: string
        origin:
          type: string
        destination:
          type: string
        baseFare:
          type: number