# lib/api-zod/src/generated/types/captain.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Tawsileh Intercity MVP API
 * OpenAPI spec version: 1.0.0
 */
import type { CaptainVehicle } from './captainVehicle';

export interface Captain {
  id: string;
  status: string;
  isOnline: boolean;
  rating?: number;
  totalTrips?: number;
  vehicle?: CaptainVehicle;
}