# lib/api-zod/src/generated/types/rideRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Tawsileh Intercity MVP API
 * OpenAPI spec version: 1.0.0
 */
import type { RideRequestCaptain } from './rideRequestCaptain';
import type { RideRequestPassenger } from './rideRequestPassenger';

export interface RideRequest {
  id: string;
  passengerId: string;
  matchedCaptainId?: string;
  routeId: string;
  status: string;
  pickupLat?: number;
  pickupLng?: number;
  destLat?: number;
  destLng?: number;
  seatsRequested?: number;
  estimatedFare: number;
  createdAt: Date;
  passenger?: RideRequestPassenger;
  captain?: RideRequestCaptain;
}