# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Tawsileh Intercity MVP API
 * OpenAPI spec version: 1.0.0
 */

export interface Route {
  id: string;
  name: string;
  origin: string;
  destination: string;
  baseFare: number;
}