# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Tawsileh Intercity MVP API
 * OpenAPI spec version: 1.0.0
 */
import type { Captain } from './captain';

export interface User {
  id: string;
  phone: string;
  fullName: string;
  role: string;
  captainProfile?: Captain;
}